refactor: folders naming
