+++ identifier/src/main/java/uz/paymo/identifier/core/utils/Exceptions.kt
@@ -2,7 +2,7 @@
  * Copyright (c) Siyavushkhon Kholmatov, 27/9/2020 at 14:4
  */
 
-package uz.paymo.identifier.core
+package uz.paymo.identifier.core.utils
 
 import android.os.Build
 import uz.paymo.identifier.BuildConfig
